Split off just the case-insensitive set operations into sorting (name WIP)
Implement initial GHI #316 sorting.py module

case_insensitive_set_intersect
case_insensitive_partial_match_set_intersect
case_insensitive_partial_match_set_diff
case_insensitive_set_diff
